+++ tests/integrations/test_sqlite.py
@@ -30,3 +30,7 @@ class DatabaseIntegrationConnectionResolver(object):
 
     def set_default_connection(self, name):
         pass
+
+    def disconnect(self):
+        if self._connection:
+            self._connection.disconnect()