+++ tests/integrations/test_mysql.py
@@ -4,26 +4,14 @@ import os
 
 from .. import OratorTestCase
 from . import IntegrationTestCase
-from orator.connections import MySQLConnection
-from orator.connectors.mysql_connector import MySQLConnector
 
 
 class MySQLIntegrationTestCase(IntegrationTestCase, OratorTestCase):
 
     @classmethod
-    def get_connection_resolver(cls):
-        return DatabaseIntegrationConnectionResolver()
-
-
-class DatabaseIntegrationConnectionResolver(object):
-
-    _connection = None
-
-    def connection(self, name=None):
-        if self._connection:
-            return self._connection
-
+    def get_manager_config(cls):
         ci = os.environ.get('CI', False)
+
         if ci:
             database = 'orator_test'
             user = 'root'
@@ -33,18 +21,12 @@ class DatabaseIntegrationConnectionResolver(object):
             user = 'orator'
             password = 'orator'
 
-        self._connection = MySQLConnection(
-            MySQLConnector().connect({
+        return {
+            'default': 'mysql',
+            'mysql': {
+                'driver': 'mysql',
                 'database': database,
                 'user': user,
                 'password': password
-            })
-        )
-
-        return self._connection
-
-    def get_default_connection(self):
-        return 'default'
-
-    def set_default_connection(self, name):
-        pass
+            }
+        }